Пробный допил для работы на 64 битном центосе. Из очевидного тока с почтой надо будет еще чет решить, остальное имх уже мелочи жизни, если на 32 битах ниче не сломалось в процессе. Ну флаги вроде работают. Правда на центосе так и не получилось добиться работы макросов из stdint для принтфа, поэтому всякие size_t пока кастуются в unsigned long (zu/zd тоже вроде не особо переносимо). Еще с морфами какая-то херня из-за шаред-поинтера, когда в конструкторе чар-даты и плеера повторно инится указатель, при сборке с оптимизацией - падает или там же на повторном ините, или позже в деструкторе, но падает стабильно. А через make test не падает, че к чему не знаю, но повторный инит закоментил, т.к. это вроде даже ничего и не меняет на выходе, да и тема уже 2 года как мертвая.

+++ src/char.cpp
@@ -1451,7 +1451,7 @@ std::string Character::get_title()
 {
 	if (!this->player_data.title) return string();
 	string tmp = string(this->player_data.title);
-	unsigned pos = tmp.find('/');
+	size_t pos = tmp.find('/');
 	if (pos == string::npos)
 		return string();
 	tmp = tmp.substr(0, pos);
@@ -1467,7 +1467,7 @@ std::string Character::get_pretitle()
 {
 	if (!this->player_data.title) return string();
 	string tmp = string(this->player_data.title);
-	unsigned pos = tmp.find('/');
+	size_t pos = tmp.find('/');
 	if (pos == string::npos)
 		return string();
 	tmp = tmp.substr(0, pos);
